CreateCROP_RECOMMENDATION_ON_ANALYZING_SOIL
India was reportedly having 60.43% of agricultural land and 58% of the population dependent on it. The farmers have enough resources though they are unable to produce up to the 
expected level of yield. One of the main reasons behind it is growing crops that don’t fulfill the requirements of soil conditions. As we know that soil is one of the most important natural resource and play a major role in crop production. Soil nutrients like nitrogen, potassium, and phosphorous are very important for any crop growth, the farmers are doing mistakes in 
choosing the crop even after having soil data that are tested by government labs. To eliminate this problem we proposed a solution i.e web application based on Artificial Intelligence and 
machine learning, which takes contents of soil as an input from a farmer and recommends an appropriate crop, the machine learning model we used is the Random forest classifier algorithm to get better accuracy. Thus, the application helps to supply information to the dilettante farmers.